# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API User Management",
    "version": "1.0.0"
  },
 "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7000/api"
    }
  ],
  "paths": {
...
    "/referentiels": {
      "post": {
        "summary": "Créer un nouveau référentiel",
        "operationId": "createReferentiel",
        "tags": ["Referentiel"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "Code unique du référentiel"
                  },
                  "libelle": {
                    "type": "string",
                    "description": "Libellé unique du référentiel"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description du référentiel"
                  },
                  "photo": {
                    "type": "string",
                    "description": "URL de la photo (facultatif)"
                  },
                  "competences": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nom": {
                          "type": "string",
                          "description": "Nom de la compétence"
                        },
                        "description": {
                          "type": "string",
                          "description": "Description de la compétence"
                        },
                        "duree_acquisition": {
                          "type": "integer",
                          "description": "Durée d'acquisition de la compétence (en jours)"
                        },
                        "modules": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "nom": {
                                "type": "string",
                                "description": "Nom du module"
                              },
                              "description": {
                                "type": "string",
                                "description": "Description du module"
                              },
                              "duree_acquisition": {
                                "type": "integer",
                                "description": "Durée d'acquisition du module (en jours)"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "required": ["code", "libelle", "description", "competences"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Référentiel créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID du référentiel créé"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Erreur de validation"
          }
        }
      },
      "get": {
        "summary": "Lister tous les référentiels",
        "operationId": "getReferentiels",
        "tags": ["Referentiel"],
        "responses": {
          "200": {
            "description": "Liste des référentiels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      },
                      "libelle": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "photo": {
                        "type": "string"
                      },
                      "competences": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "nom": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "duree_acquisition": {
                              "type": "integer"
                            },
                            "modules": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "nom": {
                                    "type": "string"
                                  },
                                  "description": {
                                    "type": "string"
                                  },
                                  "duree_acquisition": {
                                    "type": "integer"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/referentiels/{id}": {
      "get": {
        "summary": "Récupérer un référentiel par ID",
        "operationId": "getReferentielById",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Référentiel trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      },
      "patch": {
        "summary": "Mettre à jour un référentiel",
        "operationId": "updateReferentiel",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "libelle": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string"
                  },
                  "competences": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nom": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "duree_acquisition": {
                          "type": "integer"
                        },
                        "modules": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "nom": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "duree_acquisition": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Référentiel mis à jour avec succès"
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un référentiel",
        "operationId": "deleteReferentiel",
        "tags": ["Referentiel"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Référentiel supprimé avec succès"
          },
          "404": {
            "description": "Référentiel non trouvé"
          }
        }
      }
    },
...
  },
...
  }